Move more files in utils into subdirs
