+++ src/brain/githubMetrics/index.ts
@@ -1,6 +1,6 @@
 import { DISABLE_GITHUB_METRICS } from '@/config';
+import { wrapHandler } from '@/utils/misc/wrapHandler';
 import { githubEvents } from '@api/github';
-import { wrapHandler } from '@utils/wrapHandler';
 
 import { ossMetrics } from './ossMetrics';
 import { sentryMetrics } from './sentryMetrics';
