+++ src/brain/githubMetrics/ossMetrics.ts
@@ -1,6 +1,6 @@
 import { EmitterWebhookEvent } from '@octokit/webhooks';
 
-import { insertOss } from '@utils/metrics';
+import { insertOss } from '@/utils/db/metrics';
 
 /**
  * GitHub webhooks handler for OSS metrics
