+++ src/brain/issueLabelHandler/followups.ts
@@ -11,13 +11,13 @@ import {
   WAITING_FOR_PRODUCT_OWNER_LABEL,
   WAITING_FOR_SUPPORT_LABEL,
 } from '@/config';
+import { isFromABot } from '@/utils/github/isFromABot';
+import { isNotFromAnExternalOrGTMUser } from '@/utils/github/isNotFromAnExternalOrGTMUser';
+import { shouldSkip } from '@/utils/github/shouldSkip';
 import {
   calculateSLOViolationRoute,
   calculateSLOViolationTriage,
-} from '@utils/businessHours';
-import { isFromABot } from '@utils/isFromABot';
-import { isNotFromAnExternalOrGTMUser } from '@utils/isNotFromAnExternalOrGTMUser';
-import { shouldSkip } from '@utils/shouldSkip';
+} from '@/utils/misc/businessHours';
 
 function isNotInARepoWeCareAboutForFollowups(payload, org) {
   return !org.repos.all.includes(payload.repository.name);
