+++ src/brain/issueLabelHandler/triage.ts
@@ -4,10 +4,10 @@ import { EmitterWebhookEvent } from '@octokit/webhooks';
 import * as Sentry from '@sentry/node';
 
 import { GH_ORGS, WAITING_FOR_PRODUCT_OWNER_LABEL } from '@/config';
-import { isFromOutsideCollaborator } from '@/utils/isFromOutsideCollaborator';
-import { isFromABot } from '@utils/isFromABot';
-import { isNotFromAnExternalOrGTMUser } from '@utils/isNotFromAnExternalOrGTMUser';
-import { shouldSkip } from '@utils/shouldSkip';
+import { isFromABot } from '@/utils/github/isFromABot';
+import { isFromOutsideCollaborator } from '@/utils/github/isFromOutsideCollaborator';
+import { isNotFromAnExternalOrGTMUser } from '@/utils/github/isNotFromAnExternalOrGTMUser';
+import { shouldSkip } from '@/utils/github/shouldSkip';
 
 function isAlreadyUntriaged(payload) {
   return !isAlreadyTriaged(payload);
