+++ src/utils/db/updateRequiredCheck.ts
@@ -2,8 +2,7 @@ import { BuildStatus } from '@/config';
 import { SlackMessage } from '@/config/slackMessage';
 import { CheckRun } from '@/types';
 
-import { insertBuildFailure } from '../metrics';
-
+import { insertBuildFailure } from './metrics';
 import { saveSlackMessage } from './saveSlackMessage';
 import { db } from '.';
