+++ src/utils/github/getUser/index.ts
@@ -2,10 +2,10 @@ import * as Sentry from '@sentry/node';
 
 import { SLACK_PROFILE_ID_GITHUB } from '@/config';
 import { normalizeGithubUser } from '@/utils/github/normalizeGithubUser';
+import { isSentrySlackUser } from '@/utils/slack/isSentrySlackUser';
 import { bolt } from '@api/slack';
 import { db } from '@utils/db';
 import { findUser } from '@utils/db/findUser';
-import { isSentrySlackUser } from '@utils/isSentrySlackUser';
 
 type GetUserParams = Parameters<typeof findUser>[0];
