+++ src/utils/github/isNotFromAnExternalOrGTMUser.ts
@@ -1,4 +1,4 @@
-import { getOssUserType } from '@utils/getOssUserType';
+import { getOssUserType } from '@/utils/github/getOssUserType';
 
 export async function isNotFromAnExternalOrGTMUser(payload: object) {
   const type = await getOssUserType(payload);
