+++ src/utils/misc/slugizeProductArea.test.ts
