+++ src/utils/misc/slugizeProductArea.ts
