+++ src/utils/slack/getChannelsForIssue.ts
@@ -3,7 +3,7 @@ import moment from 'moment-timezone';
 
 import { PRODUCT_OWNERS_INFO } from '@/config';
 import { getTeams } from '@/utils/github/getTeams';
-import { isTimeInBusinessHours } from '@utils/businessHours';
+import { isTimeInBusinessHours } from '@/utils/misc/businessHours';
 
 export type ChannelItem = {
   channelId: string;
