+++ src/webhooks/bootstrap-dev-env/bootstrap-dev-env.test.ts
@@ -16,7 +16,7 @@ jest.mock('@google-cloud/bigquery', () => ({
 }));
 
 import { buildServer } from '@/buildServer';
-import * as db from '@utils/metrics';
+import * as db from '@/utils/db/metrics';
 
 jest.spyOn(db, 'insert');
 jest.spyOn(db, 'insertOss');
