+++ src/webhooks/bootstrap-dev-env/bootstrap-dev-env.ts
@@ -1,6 +1,6 @@
 import { FastifyRequest } from 'fastify';
 
-import { insert } from '@utils/metrics';
+import { insert } from '@/utils/db/metrics';
 
 export async function handler(
   request: FastifyRequest<{ Body: { event: string; name: string } }>
